queue/requirements.txt
azure-core==1.37.0
azure-storage-queue==12.14.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
isodate==0.7.2
pycparser==2.23
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
python-dotenv==1.2.1